update: training successful

+++ myqlanet/preprocessing/ggb.py
@@ -11,7 +11,7 @@ class GGB():
 
     def process(self):
         self.img = cv2.imread(self.path)
-        img_yuv = cv2.cvtColor(self.img, cv2.COLOR_BGR2YUV)
+        img_yuv = cv2.cvtColor(self.img.astype('uint8'), cv2.COLOR_BGR2YUV)
         img_yuv[:,:,0] = cv2.equalizeHist(img_yuv[:,:,0])
         img_output = cv2.cvtColor(img_yuv, cv2.COLOR_YUV2BGR)
         b_,g_,_ = cv2.split(img_output)
@@ -21,7 +21,7 @@ class GGB():
         return self.img
 
     def run(self, img):
-        img_yuv = cv2.cvtColor(img, cv2.COLOR_RGB2YUV)
+        img_yuv = cv2.cvtColor(img.astype('uint8'), cv2.COLOR_RGB2YUV)
         img_yuv[:,:,0] = cv2.equalizeHist(img_yuv[:,:,0])
         img_output = cv2.cvtColor(img_yuv, cv2.COLOR_YUV2RGB)
         _, g_, b_ = cv2.split(img_output)
